Explicitly cast double to int then cast int to enum

+++ src/item.cpp
@@ -431,7 +431,7 @@ bool item::is_sided() const {
 }
 
 side item::get_side() const {
-    return static_cast<side>( get_var( "lateral", static_cast<int>( side::BOTH ) ) );
+    return static_cast<side>( static_cast<int>( get_var( "lateral", static_cast<int>( side::BOTH ) ) ) );
 }
 
 bool item::set_side (side s) {
